refactor: extract endpoint logic and use proper patterns
- Move endpoint logic from payload.config.ts to dedicated handler files

- Use /api/exercises/import for both upload and lesson conversion

- Replace filesystem access with storage URL fetch (works with Blob/S3)

- Use proper Zod-validated question block structures

- Fix error responses to return HTTP status without success:true
